% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_addCovariance.R
\name{apollo_addCovariance}
\alias{apollo_addCovariance}
\title{Adds covariance matrix to Apollo model}
\usage{
apollo_addCovariance(model, apollo_inputs)
}
\arguments{
\item{model}{A model object, as returned by \link{apollo_estimate}}

\item{apollo_inputs}{List of settings. as returned by \link{apollo_validateInputs}}
}
\value{
model.
}
\description{
Receives an estimated model object, calculates its Hessian, and classical and robust covariance matrix, and returns the  
same model object, but with these additional elements.
}
