% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_addLog.R
\name{apollo_addLog}
\alias{apollo_addLog}
\title{Writes an entry to \code{apolloLog}}
\usage{
apollo_addLog(title = "", content = "", apolloLog, book = 1)
}
\arguments{
\item{title}{Character. Title of the log entry.}

\item{content}{Content of the log entry. Can be a single element or a list.
Each element will be converted to character using \code{print},
and concatenated with a line feed in between.}

\item{apolloLog}{Environment. It contains the character vectors of titles and content.}

\item{book}{Positive scalar integer. Book number inside the log (default is 1).}
}
\value{
TRUE if writing was succesful, FALSE if not.
}
\description{
Writes an entry to the apolloLog, which lives inside \code{apollo_inputs}.
}
\details{
The variable \code{apolloLog} is an environment created inside \code{apollo_inputs}
by \code{apollo_validateInputs}, but re-set by apollo_estimate.
As an environment, it can be modified in place, i.e. all changes done within this 
function are recorded in apolloLog, even if it belongs to another environment.
}
