% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_diagnostics.R
\name{apollo_diagnostics}
\alias{apollo_diagnostics}
\title{Pre-process input for multiple models
return}
\usage{
apollo_diagnostics(inputs, modelType, apollo_inputs, data = TRUE, param = TRUE)
}
\arguments{
\item{inputs}{List of settings}

\item{modelType}{Character. Type of model, e.g. "mnl", "nl", "cnl", etc.}

\item{apollo_inputs}{List of main inputs to the model estimation process. See \link{apollo_validateInputs}.}

\item{data}{Boolean. TRUE for printing report related to dependant and independant variables. FALSE for not printing it. Default is TRUE.}

\item{param}{Boolean. TRUE for printing report related to estimated parameters (e.g. model structure). FALSE for not printing it. Default is TRUE.}
}
\value{
(invisibly) TRUE if no error happend during execution.
}
\description{
Pre-process input for multiple models
return
}
