% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_initialise.R
\name{apollo_initialise}
\alias{apollo_initialise}
\title{Prepares environment}
\usage{
apollo_initialise()
}
\value{
Nothing.
}
\description{
Prepares environment (the global environment if called by the user) for model definition and estimation.
}
\details{
This function detaches variables and makes sure that output is directed to console. It does not delete variables from the working environment.
}
