% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_insertQuotes.R
\name{apollo_insertRRMQuotes}
\alias{apollo_insertRRMQuotes}
\title{Introduces quotes into rrm_settings}
\usage{
apollo_insertRRMQuotes(f)
}
\arguments{
\item{f}{Function. Usually \code{apollo_probabilities}.}
}
\value{
Function \code{f} with relevant expressions turned into character.
}
\description{
Takes a function, looks for the definition of relevant parts of rrm_settings,
and introduces quotes on them. This is to facilitate their processing by
apollo_rrm under functionality="preprocessing".
}
