% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_insertScaling.R
\name{apollo_insertScaling}
\alias{apollo_insertScaling}
\title{Scales variables inside a function}
\usage{
apollo_insertScaling(e, sca)
}
\arguments{
\item{e}{Function, expression, call or symbol to alter.}

\item{sca}{Named numeric vector with the scales. The names in these vectors determine which variables should be scaled.}
}
\value{
A function, expression, call or symbol with the corresponding variables scaled.
}
\description{
It changes the syntax of the function by replacing variable names for their scaled form, 
e.g. x --> x*apollo_inputs$apollo_scale[["x"]]. In assignments, it only scales the
right side of the assignment.
}
