% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_keepRows.R
\name{apollo_keepRows}
\alias{apollo_keepRows}
\title{Keeps only some rows}
\usage{
apollo_keepRows(v, r)
}
\arguments{
\item{v}{Numeric scalar, vector, matrix or 3-dim array.}

\item{r}{Boolean vector. As many elements as rows in \code{v}. TRUE for keeping the row. \code{FALSE} to drop it.}
}
\value{
The same argument \code{v} but with the rows where \code{r==FALSE} removed.
}
\description{
Given a numeric object (scalar, vector, matrix or 3-dim array) keeps only the specieifed rows.
}
