% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_lcConditionals.R
\name{apollo_lcConditionals}
\alias{apollo_lcConditionals}
\title{Calculates conditionals for latent class models.}
\usage{
apollo_lcConditionals(model, apollo_probabilities, apollo_inputs)
}
\arguments{
\item{model}{Model object. Estimated model object as returned by function \link{apollo_estimate}.}

\item{apollo_probabilities}{Function. Returns probabilities of the model to be estimated. Must receive three arguments:
\itemize{
  \item \strong{\code{apollo_beta}}: Named numeric vector. Names and values of model parameters.
  \item \strong{\code{apollo_inputs}}: List containing options of the model. See \link{apollo_validateInputs}.
  \item \strong{\code{functionality}}: Character. Can be either \strong{\code{"components"}}, \strong{\code{"conditionals"}}, \strong{\code{"estimate"}} (default), \strong{\code{"gradient"}}, \strong{\code{"output"}}, \strong{\code{"prediction"}}, \strong{\code{"preprocess"}}, \strong{\code{"raw"}}, \strong{\code{"report"}}, \strong{\code{"shares_LL"}}, \strong{\code{"validate"}} or \strong{\code{"zero_LL"}}.
}}

\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}
}
\value{
A matrix with the posterior class allocation probabilities for each individual.
}
\description{
Calculates posterior expected values (conditionals) of class allocation probabilities for each individual.
}
\details{
This function can only be used with latent class models without continuous heterogeneity.
}
