% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_lcUnconditionals.r
\name{apollo_lcUnconditionals}
\alias{apollo_lcUnconditionals}
\title{Returns draws for random parameters in a latent class model model}
\usage{
apollo_lcUnconditionals(model, apollo_probabilities, apollo_inputs)
}
\arguments{
\item{model}{Model object. Estimated model object as returned by function \link{apollo_estimate}.}

\item{apollo_probabilities}{Function. Returns probabilities of the model to be estimated. Must receive three arguments:
\itemize{
  \item apollo_beta: Named numeric vector. Names and values of model parameters.
  \item apollo_inputs: List containing options of the model. See \link{apollo_validateInputs}.
  \item functionality: Character. Can be either "estimate" (default), "prediction", "validate", "conditionals", "zero_LL", or "raw".
}}

\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}
}
\value{
List of object, one per random component and one for the class allocation probabilities.
}
\description{
Returns draws (unconditionals) for random parameters in model, including interactions with deterministic covariates
}
\details{
This functions is only meant for use with continuous distributions
}
