% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_loadModel.R
\name{apollo_loadModel}
\alias{apollo_loadModel}
\title{Loads model from file}
\usage{
apollo_loadModel(modelName)
}
\arguments{
\item{modelName}{Character. Name of the model to load.}
}
\value{
A model object.
}
\description{
Loads a previously estimated model object from a file.
}
\details{
This function looks for a file named \code{modelName_model.rds} in the working or output directory, loads the object contained in it, and returns it.
}
