% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_lrTest.R
\name{apollo_lrTest}
\alias{apollo_lrTest}
\title{Likelihood ratio test}
\usage{
apollo_lrTest(baseModel, generalModel)
}
\arguments{
\item{baseModel}{Character. Name of a previously estimated model whose results were written to disk by \link{apollo_saveOutput}.
This is the restricted model, i.e. the one with fewer parameters.}

\item{generalModel}{Either a character variable with the name of a previously estimated model, or an estimated model in memory, as
returned by \link{apollo_estimate}. This model nests \code{baseModel}, and it should have more parameters than it.}
}
\value{
LL ratio test statistic (invisibly)
}
\description{
Calculates the likelihood ratio test and prints result.
}
