% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_mdcevOutside.R
\name{apollo_mdcevOutside}
\alias{apollo_mdcevOutside}
\title{Calculates MDCEV likelihoods with an outside good.}
\usage{
apollo_mdcevOutside(
  V,
  alternatives,
  alpha,
  gamma,
  sigma,
  cost,
  avail,
  continuousChoice,
  budget,
  functionality,
  minConsumption = NA,
  outsideName = "outside",
  rows = "all",
  componentName = "MDCEV"
)
}
\arguments{
\item{V}{Named list. Utilities of the alternatives. Names of elements must match those in argument 'alternatives'.}

\item{alternatives}{Character vector. Names of alternatives, elements must match the names in list 'V'.}

\item{alpha}{Named list. Alpha parameters for each alternative, including for the outside good. As many elements as alternatives.}

\item{gamma}{Named list. Gamma parameters for each alternative, excluding the outside good. As many elements as inside good alternatives.}

\item{sigma}{Numeric scalar. Scale parameter of the model extreme value type I error.}

\item{cost}{Named list of numeric vectors. Price of each alternative. One element per alternative, each one as long as the number of observations or a scalar. Names must match those in \code{alternatives}.}

\item{avail}{Named list. Availabilities of alternatives, one element per alternative. Names of elements must match those in argument 'alternatives'. Value for each element can be 1 (scalar if always available) or a vector with values 0 or 1 for each observation. If all alternatives are always available, then user can just omit this argument.}

\item{continuousChoice}{Named list of numeric vectors. Amount of consumption of each alternative. One element per alternative, as long as the number of observations or a scalar. Names must match those in \code{alternatives}.}

\item{budget}{Numeric vector. Budget for each observation.}

\item{functionality}{Character. Can take different values depending on desired output.
\itemize{
  \item "estimate" Used for model estimation.
  \item "prediction" Used for model predictions.
  \item "validate" Used for validating input.
  \item "zero_LL" Used for calculating null likelihood.
  \item "conditionals" Used for calculating conditionals.
  \item "output" Used for preparing output after model estimation.
  \item "raw" Used for debugging.
}}

\item{minConsumption}{Named list of scalars or numeric vectors. Minimum consumption of the alternatives, if consumed. As many elements as alternatives. Names must match those in \code{alternatives}.}

\item{outsideName}{Character. Alternative name for the outside good. The first good is assumed to be the outside one. Default is "outside"}

\item{rows}{Boolean vector. Consideration of rows in the likelihood calculation, FALSE to exclude. Length equal to the number of observations (nObs). Default is \code{"all"}, equivalent to \code{rep(TRUE, nObs)}.}

\item{componentName}{Character. Name given to model component.}
}
\value{
The returned object depends on the value of argument \code{functionality} as follows.
        \itemize{
          \item "estimate": vector/matrix/array. Returns the probabilities for the chosen alternative for each observation.
          \item "prediction": A matrix with one row per observation, and means and s.d. of predicted consumptions.
          \item "validate": Boolean. Returns TRUE if all tests are passed.
          \item "zero_LL": Not applicable.
          \item "conditionals": Same as "prediction".
          \item "output": Same as "estimate" but also writes summary of choices into temporary file (later read by \code{apollo_modelOutput}).
          \item "raw": Same as "prediction".
        }
}
\description{
Calculates the likelihood of a Multiple Discrete Continuous Extreme Value (MDCEV) model with an outside good.
}
