% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_mlhs.R
\name{apollo_mlhs}
\alias{apollo_mlhs}
\title{Generate random draws using MLHS algorithm}
\usage{
apollo_mlhs(N, d, i)
}
\arguments{
\item{N}{The number of draws to generate in each dimension}

\item{d}{The number of dimensions to generate draws in}

\item{i}{The number of individuals to generate draws for}
}
\value{
A (N*i) x d matrix with random draws
}
\description{
Generate random draws using the Modified Latin Hypercube Sampling algorithm.
}
\details{
Internal use only.
Algorithm described in
Hess, S., Train, K., and Polak, J. (2006) Transportation Research 40B, 147 - 163.
}
