% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_print.R
\name{apollo_print}
\alias{apollo_print}
\title{Prints message to terminal}
\usage{
apollo_print(txt, nSignifD = 4, widthLim = 11, pause = 0, type = "t")
}
\arguments{
\item{txt}{Character, what to print.}

\item{nSignifD}{Optional numeric integer. Minimum number of significant digits when printing numeric matrices. Default is 4.}

\item{widthLim}{Optional numeric integer. Minimum width (in characters) of each column when printing numeric matrices. Default is 11}

\item{pause}{Scalar integer. Number of seconds the execution will pause after printing the message. Default is 0.}

\item{type}{Character. "t" for regular text (default), "w" for warning, "i" for information.}
}
\value{
Nothing
}
\description{
Prints message to terminal if \code{apollo_inputs$silent} is FALSE
}
