% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_reportModelTypeLog.R
\name{apollo_reportModelTypeLog}
\alias{apollo_reportModelTypeLog}
\title{Writes the type of a model inside \code{apolloLog}}
\usage{
apollo_reportModelTypeLog(modelType, apolloLog)
}
\arguments{
\item{modelType}{Character. Type of the model, e.g. "MNL" or "MDCEV"}

\item{apolloLog}{Environment. It contains the character vectors of titles and content.}
}
\value{
TRUE if writing was succesful, FALSE if not.
}
\description{
Writes the type of the model in a list inside apolloLog, which lives inside \code{apollo_inputs}.
}
\details{
The variable \code{apolloLog} is an environment created inside \code{apollo_inputs}
by \code{apollo_validateInputs}, but re-set by apollo_estimate.
As an environment, it can be modified in place, i.e. all changes done within this 
function are recorded in apolloLog, even if it belongs to another environment.
}
