% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_setWorkDir.R
\name{apollo_setWorkDir}
\alias{apollo_setWorkDir}
\title{Automatically sets working directory to active file directory}
\usage{
apollo_setWorkDir()
}
\value{
(invisibly) TRUE if it manages to set the working directory, 
        FALSE if not.
}
\description{
This function only works in Rstudio. If called outside RStudio
will just print a message to screen saying it could not set the 
working directory.
}
