% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_splitDataDraws.R
\name{apollo_splitDataDraws}
\alias{apollo_splitDataDraws}
\title{Splits data and draws for loading in cluster}
\usage{
apollo_splitDataDraws(apollo_inputs, silent = FALSE)
}
\arguments{
\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}

\item{silent}{Boolean. If TRUE, no information is printed to console or default output.}
}
\value{
List. Each element is a copy of apollo_inputs, but with only a piece of \code{database} and \code{draws}.
}
\description{
Copies \code{apollo_inputs} as many times as cores, but each copy contains only part of \code{database} and \code{draws}.
}
\details{
Internal use only. This function is called by \link{apollo_makeCluster}.
}
