% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_validateData.R
\name{apollo_validateData}
\alias{apollo_validateData}
\title{Validates data}
\usage{
apollo_validateData(database, apollo_control, silent)
}
\arguments{
\item{database}{data.frame. Data used by model.}

\item{apollo_control}{List. Options controlling the running of the code.
See \code{?apollo_validateControl} for details.}

\item{silent}{Boolean. TRUE to keep the function from printing to the console.
Default is FALSE.}
}
\value{
Data.frame. Validated version of database.
}
\description{
Checks consistency of the database with \code{apollo_control}, sorts it by indivID, and adds an internal ID variable (\code{apollo_sequence})
}
\details{
This function should be called after calling apollo_validateControl.
Observations are sorted only if apollo_control$panelData=TRUE.
}
