% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_varList.R
\name{apollo_varList}
\alias{apollo_varList}
\title{Lists variables names and definitions used in a model}
\usage{
apollo_varList(apollo_probabilities, apollo_inputs)
}
\arguments{
\item{apollo_probabilities}{Likelihood function of the whole model.}

\item{apollo_inputs}{List of arguments and settings generated by \link{apollo_validateInputs}.}
}
\value{
A list of expressions containing all definitions in apollo_randCoeff and apollo_probabilities
}
\description{
Returns a list containing the names and definitions of variables in apollo_probabilities, apollo_randCoeff and apollo_lcPars
}
\details{
It looks for variables in apollo_beta, apollo_randCoeff, draws, and 
apollo_probabilities. It returns them in a list ordered by origin.
}
