% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onAttach.R
\name{.onAttach}
\alias{.onAttach}
\title{Prints package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Name of library.}

\item{pkgname}{Name of package.}
}
\value{
Nothing
}
\description{
This function is only called by R when attaching the package.
}
