% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.apollo.R
\name{print.apollo}
\alias{print.apollo}
\title{Prints brief summary of Apollo model}
\usage{
\method{print}{apollo}(x, ...)
}
\arguments{
\item{x}{Model object. Estimated model object as returned by function \link{apollo_estimate}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
nothing.
}
\description{
Receives an estimated model object and prints a brief summary using the generic print function.
}
