% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_checkArguments.R
\name{apollo_checkArguments}
\alias{apollo_checkArguments}
\title{Checks definitions Apollo functions}
\usage{
apollo_checkArguments(
  apollo_probabilities = NA,
  apollo_randCoeff = NA,
  apollo_lcPars = NA
)
}
\arguments{
\item{apollo_probabilities}{Function. Likelihood function as defined by the user.}

\item{apollo_randCoeff}{Function. Defines the random components used inside \code{apollo_probabilities}.}

\item{apollo_lcPars}{Function. Defines the class allocation probabilities and (optionally) the lists of
parameters of each class for latent class models.}
}
\value{
Returns (invisibly) TRUE if definitions are correct, and FALSE otherwise.
}
\description{
Checks that the user-defined functions used by Apollo are correctly defined by the user.
}
\details{
It only checks that the functions have the correct definition of inputs. It does not run the functions.
}
