% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_conditionals.R
\name{apollo_conditionals}
\alias{apollo_conditionals}
\title{Calculates conditionals}
\usage{
apollo_conditionals(model, apollo_probabilities, apollo_inputs)
}
\arguments{
\item{model}{Model object. Estimated model object as returned by function \link{apollo_estimate}.}

\item{apollo_probabilities}{Function. Returns probabilities of the model to be estimated. Must receive three arguments:
\itemize{
  \item apollo_beta: Named numeric vector. Names and values of model parameters.
  \item apollo_inputs: List containing options of the model. See \link{apollo_validateInputs}.
  \item functionality: Character. Can be either "estimate" (default), "prediction", "validate", "conditionals", "zero_LL", or "raw".
}}

\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}
}
\value{
List of matrices. Each matrix has dimensions nIndiv x 3. One matrix per random component.
        Each row of each matrix contains the indivID of an individual, and the
        posterior mean and s.d. of this random component for this individual
}
\description{
Calculates posterior expected values (conditionals) of random coefficients, as well as their standard deviations.
}
\details{
This functions is only meant for use with continuous distributions
}
