% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_lrTest.R
\name{apollo_lrTest}
\alias{apollo_lrTest}
\title{Likelihood ratio test}
\usage{
apollo_lrTest(model1, model2)
}
\arguments{
\item{model1}{Either a character variable with the name of a previously estimated model, or an estimated 
model in memory, as returned by \link{apollo_estimate}.}

\item{model2}{Either a character variable with the name of a previously estimated model, or an estimated 
model in memory, as returned by \link{apollo_estimate}.}
}
\value{
LR-test p-value (invisibly)
}
\description{
Calculates the likelihood ratio test value between two models and reports the corresponding p-value. 
The two models need to have been estimated on the same data, and one model needs to be nested within the other model.
}
