% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_makeLogLike.R
\name{apollo_makeLogLike}
\alias{apollo_makeLogLike}
\title{Creates loglikelihood function.}
\usage{
apollo_makeLogLike(theta_fix_val, database, apollo_probabilities,
  apollo_control, draws = NA, apollo_randcoeff = NA,
  apollo_lcpars = NA, cl = NA, estimation_routine, work_in_logs)
}
\arguments{
\item{theta_fix_val}{Named numeric vector. Names an values of fixed parameters.}

\item{database}{data.frame Data to be used by model. Gets splitted and copied into the cluster.}

\item{apollo_probabilities}{Function. likelihood provided by the user.}

\item{apollo_control}{List of model options.}

\item{draws}{List of draws (numeric arrays). Default is NA. Get splitted and copied into the cluster.}

\item{apollo_randcoeff}{Function. Creates the random coeff or error terms by mixing draws and params. Provided by the user.}

\item{apollo_lcpars}{Function. Used for latent class models. Not implemented yet.}

\item{cl}{Cluster.}

\item{estimation_routine}{Character.}

\item{work_in_logs}{Boolean.}
}
\value{
apollo_logLike function.
}
\description{
Creates loglikelihood function from the likelihood function apollo_probabilities provided by the user.
}
\details{
Internal use only. Called by apollo_estimate before estimation.
}
