% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_mdcnev.R
\name{apollo_mdcnev}
\alias{apollo_mdcnev}
\title{Calculates MDCNEV likelihoods with an outside good.}
\usage{
apollo_mdcnev(mdcnev_settings, functionality)
}
\arguments{
\item{mdcnev_settings}{List of settings for the MDCEV model. It must include the following.
\itemize{
  \item V: Named list. Utilities of the alternatives. Names of elements must match those in argument 'alternatives'.
  \item alternatives: Character vector. Names of alternatives, elements must match the names in list 'V'.
  \item alpha: Named list. Alpha parameters for each alternative, including for the outside good. As many elements as alternatives.
  \item gamma: Named list. Gamma parameters for each alternative, including for the outside good. As many elements as alternatives.
  \item mdcnevNests: Named list. Lambda parameters for each nest. Elements must be named with the nest name. The lambda at the root is fixed to 1, and therefore must be no be defined. The value of the estimated mdcnevNests parameters should be between 0 and 1 to ensure consistency with random utility maximization.
  \item mdcnevStructure: Numeric matrix. One row per nest and one column per alternative. Each element of the matrix is 1 if an alternative belongs to the corresponding nest.
  \item cost: Named list of numeric vectors. Price of each alternative. One element per alternative, each one as long as the number of observations or a scalar. Names must match those in \code{alternatives}.
  \item avail: Named list. Availabilities of alternatives, one element per alternative. Names of elements must match those in argument 'alternatives'. Value for each element can be 1 (scalar if always available) or a vector with values 0 or 1 for each observation. If all alternatives are always available, then user can just omit this argument.
  \item continuousChoice: Named list of numeric vectors. Amount of consumption of each alternative. One element per alternative, as long as the number of observations or a scalar. Names must match those in \code{alternatives}.
  \item budget: Numeric vector. Budget for each observation.
  \item minConsumption: Named list of scalars or numeric vectors. Minimum consumption of the alternatives, if consumed. As many elements as alternatives. Names must match those in \code{alternatives}.
  \item outside: Character. Alternative name for the outside good. Default is "outside"
  \item rows: Boolean vector. Consideration of rows in the likelihood calculation, FALSE to exclude. Length equal to the number of observations (nObs). Default is \code{"all"}, equivalent to \code{rep(TRUE, nObs)}.
}}

\item{functionality}{Character. Can take different values depending on desired output.
\itemize{
  \item "estimate" Used for model estimation.
  \item "prediction" Used for model predictions.
  \item "validate" Used for validating input.
  \item "zero_LL" Used for calculating null likelihood.
  \item "conditionals" Used for calculating conditionals.
  \item "output" Used for preparing output after model estimation.
  \item "raw" Used for debugging.
}}
}
\value{
The returned object depends on the value of argument \code{functionality} as follows.
        \itemize{
          \item "estimate": vector/matrix/array. Returns the probabilities for the chosen alternative for each observation.
          \item "prediction": A matrix with one row per observation, and means and s.d. of predicted consumptions.
          \item "validate": Boolean. Returns TRUE if all tests are passed.
          \item "zero_LL": Not applicable.
          \item "conditionals": Same as "prediction".
          \item "output": Same as "estimate" but also writes summary of choices into temporary file (later read by \code{apollo_modelOutput}).
          \item "raw": Same as "prediction".
        }
}
\description{
Calculates the likelihood of a Multiple Discrete Continuous Nested Extreme Value (MDCNEV) model with an outside good.
}
