% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_preprocess.R
\name{apollo_preprocess}
\alias{apollo_preprocess}
\title{Pre-process input for multiple models
return}
\usage{
apollo_preprocess(inputs, modelType, functionality, apollo_inputs)
}
\arguments{
\item{inputs}{List of settings}

\item{modelType}{Character. Type of model, e.g. "mnl", "nl", "cnl", etc.}

\item{functionality}{Character. Either "estimate","prediction","validate","zero_LL","conditionals","output","raw", or "preprocess". Only used for validation, it does not influence the return values.}

\item{apollo_inputs}{List of main inputs to the model estimation process. See \link{apollo_validateInputs}.}
}
\value{
The returned object depends on the value of argument operation
}
\description{
Pre-process input for multiple models
return
}
