% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_printLog.R
\name{apollo_printLog}
\alias{apollo_printLog}
\title{Returns the log of Apollo}
\usage{
apollo_printLog(apolloLog)
}
\arguments{
\item{apolloLog}{Environment. It contains the character vectors of titles and content.}
}
\value{
A list or a scalar character variable.
}
\description{
Returns the apolloLog variable either as a list or as text.
}
\details{
The variable apolloLog is a list whose elements are character vectors with two elements.
The first element is the title of the entry, and the second element is content of the entry.
ApolloLog lives in the namespace environment of the Apollo package.
}
