% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_splitDataDraws.R
\name{apollo_splitDataDraws}
\alias{apollo_splitDataDraws}
\title{Splits data and draws for loading in cluster}
\usage{
apollo_splitDataDraws(apollo_inputs, silent = FALSE)
}
\arguments{
\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}

\item{silent}{Boolean. If TRUE, no information is printed to console or default output.}
}
\value{
Character vector of file names of the pieces of apollo_inputs.
}
\description{
Splits apollo_inputs into pieces and writes them to disk (temporary folder).
}
\details{
Internal use only. This function is called by \link{apollo_makeCluster}.
}
