% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_validateControl.R
\name{apollo_validateControl}
\alias{apollo_validateControl}
\title{Validates apollo_control}
\usage{
apollo_validateControl(database, apollo_control, silent = FALSE)
}
\arguments{
\item{database}{data.frame. Data used by model.}

\item{apollo_control}{List. Options controlling the running of the code. User input is required for all settings except those with a default or marked as optional. 
\itemize{
  \item \strong{\code{calculateLLC}}: Boolean. TRUE if user wants to calculate LL at constants (if applicable). - TRUE by default.
  \item \strong{\code{HB}}: Boolean. TRUE if using RSGHB for Bayesian estimation of model.
  \item \strong{\code{indivID}}: Character. Name of column in the database with each decision maker's ID.
  \item \strong{\code{memorySaver}}: Boolean. TRUE to reduce memory usage when calculating analytical gradients and hessian - FALSE by default.
  \item \strong{\code{mixing}}: Boolean. TRUE for models that include random parameters.
  \item \strong{\code{modelDescr}}: Character. Description of the model. Used in output files.
  \item \strong{\code{modelName}}: Character. Name of the model. Used when saving the output to files.
  \item \strong{\code{nCores}}: Numeric>0. Number of cores to use in calculations of the model likelihood.
  \item \strong{\code{noDiagnostics}}: Boolean. TRUE if user does not wish model diagnostics to be printed - FALSE by default.
  \item \strong{\code{noValidation}}: Boolean. TRUE if user does not wish model input to be validated before estimation - FALSE by default.
  \item \strong{\code{outputDirectory}}: Character. Optional directory for outputs if different from working director - empty by default
  \item \strong{\code{panelData}}: Boolean. TRUE if there are multiple observations (i.e. rows) for each decision maker - Automatically set based on \code{indivID} by default.
  \item \strong{\code{seed}}: Numeric. Seed for random number generation.
  \item \strong{\code{weights}}: Character. Name of column in database containing weights for estimation.
  \item \strong{\code{workInLogs}}: Boolean. TRUE for increased numeric precision in models with panel data - FALSE by default.
}}

\item{silent}{Boolean. If TRUE, no messages are printed to screen.}
}
\value{
Validated version of \code{apollo_control}, with additional element called \code{panelData} set to TRUE for repeated choice data.
}
\description{
Validates the options controlling the running of the code \code{apollo_control} and sets default values for the omitted ones.
}
\details{
This function should be run before running \code{apollo_validateData}.
}
