% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_validatedata.R
\name{apollo_validatedata}
\alias{apollo_validatedata}
\title{Validates data}
\usage{
apollo_validatedata(database, apollo_control)
}
\arguments{
\item{database}{data.frame. Data used by model.}

\item{apollo_control}{List. Options controlling the running of the code.
See \code{?apollo_validatecontrol} for details.}
}
\value{
Data.frame. Validated version of database.
}
\description{
Checks consistency of the database with \code{apollo_control}, orders it by indivID, and adds an internal ID variable (\code{apollo_ID})
}
\details{
This function should be called after calling apollo_validatecontrol.
}
