% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_varList.R
\name{apollo_varList}
\alias{apollo_varList}
\title{Lists variable names and definitions used inside a function}
\usage{
apollo_varList(f, apollo_inputs)
}
\arguments{
\item{f}{A function, usually \code{apollo_probabilities}}

\item{apollo_inputs}{List grouping most common inputs. Created by function \link{apollo_validateInputs}.}
}
\value{
A list of expressions containing all definitions in f, \code{apollo_randCoeff} and \code{apollo_probabilities}
}
\description{
Returns a list containing the names and definitions of variables in f, \code{apollo_randCoeff} and \code{apollo_lcPars}
}
\details{
It looks for variable definitions inside f, \code{apollo_randCoeff}, and \code{apollo_lcPars}. It returns 
them in a list.
}
