% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apollo_writeF12.R
\name{apollo_writeF12}
\alias{apollo_writeF12}
\title{Writes an F12 file}
\usage{
apollo_writeF12(model, truncateCoeffNames = TRUE)
}
\arguments{
\item{model}{Model object. Estimated model object as returned by function \link{apollo_estimate}.}

\item{truncateCoeffNames}{Boolean. TRUE to truncate parameter names to 10 characters.
TRUE by default.}
}
\value{
Nothing.
}
\description{
Writes an F12 file (ALogit format) with the results of a model estimation.
}
